+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { delayExecution, type CancelToken } from './delayExecution';
+export { throttle } from './throttle';
